+++ src/core/reducers/api.js
@@ -2,23 +2,15 @@
 import UAParser from 'ua-parser-js';
 
 import { LOG_OUT_USER } from 'amo/reducers/users';
-import {
-  SET_AUTH_TOKEN,
-  SET_LANG,
-  SET_CLIENT_APP,
-  SET_REGION_CODE,
-  SET_REQUEST_ID,
-  SET_USER_AGENT,
-} from 'core/constants';
 import type { LogOutUserAction } from 'amo/reducers/users';
-import type {
-  SetAuthTokenAction,
-  SetClientAppAction,
-  SetLangAction,
-  SetRequestIdAction,
-  SetUserAgentAction,
-} from 'core/actions/index';
 import type { Exact } from 'core/types/util';
+
+export const SET_AUTH_TOKEN = 'SET_AUTH_TOKEN';
+export const SET_CLIENT_APP = 'SET_CLIENT_APP';
+export const SET_LANG = 'SET_LANG';
+export const SET_REGION_CODE = 'SET_REGION_CODE';
+export const SET_REQUEST_ID = 'SET_REQUEST_ID';
+export const SET_USER_AGENT = 'SET_USER_AGENT';
 
 export const USER_AGENT_BROWSER_FIREFOX = 'Firefox';
 export const USER_AGENT_OS_ANDROID: 'Android' = 'Android';
@@ -87,6 +79,84 @@ export const initialApiState: ApiState = {
   userAgent: null,
   userAgentInfo: { browser: {}, os: {} },
 };
+
+export type SetAuthTokenAction = {|
+  payload: {| token: string |},
+  type: typeof SET_AUTH_TOKEN,
+|};
+
+export function setAuthToken(token: string): SetAuthTokenAction {
+  if (!token) {
+    throw new Error('token cannot be falsey');
+  }
+  return {
+    type: SET_AUTH_TOKEN,
+    payload: { token },
+  };
+}
+
+export type SetClientAppAction = {|
+  payload: {| clientApp: string |},
+  type: typeof SET_CLIENT_APP,
+|};
+
+export function setClientApp(clientApp: string): SetClientAppAction {
+  if (!clientApp) {
+    throw new Error('clientApp cannot be falsey');
+  }
+  return {
+    type: SET_CLIENT_APP,
+    payload: { clientApp },
+  };
+}
+
+export type SetLangAction = {|
+  payload: {| lang: string |},
+  type: typeof SET_LANG,
+|};
+
+export function setLang(lang: string): SetLangAction {
+  return {
+    type: SET_LANG,
+    payload: { lang },
+  };
+}
+
+export type SetRegionCodeAction = {|
+  payload: {| regionCode: string |},
+  type: typeof SET_REGION_CODE,
+|};
+
+export function setRegionCode(regionCode: string): SetRegionCodeAction {
+  return {
+    type: SET_REGION_CODE,
+    payload: { regionCode },
+  };
+}
+
+export type SetUserAgentAction = {|
+  payload: {| userAgent: string |},
+  type: typeof SET_USER_AGENT,
+|};
+
+export function setUserAgent(userAgent: string): SetUserAgentAction {
+  return {
+    type: SET_USER_AGENT,
+    payload: { userAgent },
+  };
+}
+
+export type SetRequestIdAction = {|
+  payload: {| requestId: string |},
+  type: typeof SET_REQUEST_ID,
+|};
+
+export function setRequestId(requestId: string): SetRequestIdAction {
+  return {
+    type: SET_REQUEST_ID,
+    payload: { requestId },
+  };
+}
 
 type Action =
   | SetAuthTokenAction
