+++ src/core/server/base.js
@@ -19,9 +19,9 @@ import WebpackIsomorphicTools from 'webpack-isomorphic-tools';
 import { ChunkExtractor, ChunkExtractorManager } from '@loadable/server';
 
 import log from 'core/logger';
-import { createApiError } from 'core/api';
+import { REGION_CODE_HEADER, createApiError } from 'core/api';
 import Root from 'core/components/Root';
-import { AMO_REQUEST_ID_HEADER, REGION_CODE_HEADER } from 'core/constants';
+import { AMO_REQUEST_ID_HEADER } from 'core/constants';
 import ServerHtml from 'core/components/ServerHtml';
 import * as middleware from 'core/middleware';
 import requestId from 'core/middleware/requestId';
@@ -36,7 +36,7 @@ import {
   setRegionCode,
   setRequestId,
   setUserAgent,
-} from 'core/actions';
+} from 'core/reducers/api';
 import {
   getDirection,
   isValidLang,
@@ -92,8 +92,8 @@ export function getPageProps({ store, req, res, config }) {
     log.debug('No userAgent found in request headers.');
   }
 
-  if (req.headers[REGION_CODE_HEADER]) {
-    store.dispatch(setRegionCode(req.headers[REGION_CODE_HEADER]));
+  if (req.get(REGION_CODE_HEADER)) {
+    store.dispatch(setRegionCode(req.get(REGION_CODE_HEADER)));
   } else {
     log.debug(`No ${REGION_CODE_HEADER} found in request headers.`);
   }
