+++ tests/unit/amo/components/TestApp.js
@@ -15,7 +15,7 @@ import NotFoundPage from 'amo/pages/ErrorPages/NotFoundPage';
 import ServerErrorPage from 'amo/pages/ErrorPages/ServerErrorPage';
 import { logOutUser as logOutUserAction } from 'amo/reducers/users';
 import createStore from 'amo/store';
-import { setUserAgent as setUserAgentAction } from 'core/actions';
+import { setUserAgent as setUserAgentAction } from 'core/reducers/api';
 import {
   CLIENT_APP_ANDROID,
   CLIENT_APP_FIREFOX,
