+++ tests/unit/amo/components/TestSectionLinks.js
@@ -2,7 +2,7 @@ import * as React from 'react';
 
 import { setViewContext } from 'amo/actions/viewContext';
 import SectionLinks, { SectionLinksBase } from 'amo/components/SectionLinks';
-import { setClientApp } from 'core/actions';
+import { setClientApp } from 'core/reducers/api';
 import {
   ADDON_TYPE_EXTENSION,
   ADDON_TYPE_STATIC_THEME,
