+++ tests/unit/core/server/test_base.js
@@ -11,12 +11,11 @@ import supertest from 'supertest';
 import defaultConfig from 'config';
 import cheerio from 'cheerio';
 
-import { setRegionCode, setRequestId } from 'core/actions';
-import { createApiError } from 'core/api';
-import { AMO_REQUEST_ID_HEADER, REGION_CODE_HEADER } from 'core/constants';
+import { REGION_CODE_HEADER, createApiError } from 'core/api';
+import { AMO_REQUEST_ID_HEADER } from 'core/constants';
 import baseServer, { createHistory } from 'core/server/base';
 import { middleware } from 'core/store';
-import apiReducer from 'core/reducers/api';
+import apiReducer, { setRegionCode, setRequestId } from 'core/reducers/api';
 import redirectToReducer, {
   sendServerRedirect,
 } from 'core/reducers/redirectTo';
